api-gateway/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
python-jose[cryptography]==3.3.0
redis==5.0.1
tenacity==8.2.3
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
prometheus-fastapi-instrumentator==6.1.0
